use get_or_create for shorten_url retries
If the initial get_or_create fails in shorten_url, we enter
an infinite loop of create() attempts. If the initial failure
was due to competition with another instance of trying to
shorten the same URL, then the subsequent create() call will
never succeed and the code will spin until killed.

This commit changes the retries to use get_or_create as well,
so that the retry will not try to create a short url that
already exists.